Create Section_4_RadiationComponent.md
In our unified framework, Radiation can be described by 4) R=M/r^2​, representing the spatial extension of energy emanating from mass. By defining Radiation in terms of mass and distance, we bridge the gap between gravitational forces and radiative phenomena, suggesting that Radiation and Gravitation are two aspects of the same fundamental Energy Equation.

This approach allows us to describe various astrophysical phenomena in terms of the balance between Gravitation and Radiation. For instance, in environments with extreme Radiation and limited Gravitation, , such as Gamma Ray Bursts and Supernovae, we see energetic outflows that can be modeled within this framework. Conversely, in environments where Gravitation dominates and Radiation is limited, , such as Black Holes, the model predicts intense gravitational fields with minimal radiative escape. It would seem, however, that most systems are perfectly in balance, when observed from the correct Scale.

“Scale refers to the relative size of any Radiation Source.” -Definition 6 of The Unified Theory of Energy